requirements.txt
flask==3.0.0
flask-sqlalchemy==3.1.1
flask-migrate==4.0.5
flask-cors==4.0.0
python-dotenv==1.0.0
psycopg2-binary==2.9.9
openai==1.55.3  # Pin specific version for compatibility
httpx==0.27.2  # Pin specific version for compatibility with OpenAI
requests==2.31.0
gunicorn==21.2.0
python-jose==3.3.0
